feat: use markdown form for file render
